make pipeline work with cleaning pipeline changes. Finalize Database side matching step
