+++ pipeline/Dockerfile
@@ -19,6 +19,9 @@ RUN pip install --no-cache-dir -r requirements.txt
 COPY . .
 
 # Run the development server
-# TODO: trying to track down error in production cloud run
-# CMD ["bash", "setup.sh", "--migrate", "--load-geos", "--load-programs", "--load-geo-metrics"]
-CMD ["bash", "setup.sh", "--migrate", "--load-base-tables", "--load-dependent-tables", "--load-assoc-table"]
+# CMD ["bash", "setup.sh", "--migrate", "--load-base-tables", "--load-dependent-tables"]
+CMD ["bash", "setup.sh", "--migrate", "--load-assoc-table"]
+
+    # "--load-base-tables", 
+    # "--load-dependent-tables"]
+# , "--load-assoc-table"]
