+++ pipeline/tax_credit/database_helper.py
@@ -1,14 +1,17 @@
 from itertools import islice
 
+from django.db.utils import ProgrammingError
+
+from tax_credit.load_job import LoadJob
+from common.storage import DataReader
 from common.logger import LoggerFactory
-from django.db import models
 
 logger = LoggerFactory.get(__name__)
 
 
 class DatabaseHelper:
     @staticmethod
-    def load_batched(job, reader, load_batch_size, batch_number_of):
+    def load_batched(job: LoadJob, reader: DataReader, load_batch_size, batch_number_of):
         try:
             objs = (
                 job.row_to_model(row)
@@ -20,21 +23,26 @@ class DatabaseHelper:
                 batch = list(islice(objs, load_batch_size))
                 if not batch:
                     break
-                job.model.objects.bulk_create(
-                    batch,
-                    update_conflicts=True,
-                    unique_fields=job.unique_fields,
-                    update_fields=job.update_fields,
-                )
-
+                if job.unique_fields:
+                    job.model.objects.bulk_create(
+                        batch,
+                        update_conflicts=True,
+                        unique_fields=job.unique_fields,
+                        update_fields=job.update_fields,
+                    )
+                else: 
+                    job.model.objects.bulk_create(
+                        batch,
+                        update_fields=job.update_fields,
+                    )
                 batch_ct += 1
                 if batch_number_of is not None and batch_ct >= batch_number_of:
                     break
-        except Exception as e:
+        except ProgrammingError as e:
             logger.error(
-                f"ERROR!!!!! Could not process [ {job.model} ] record. Check row to model transformation."
+                f"ERROR!!!!! Could not process [ {job.job_name} ] record. Check row to model transformation: [ {job.row_to_model} ]"
             )
-            logger.error(f"ERROR : {e}")
+            # logger.error(f"ERROR : {e}")
 
     @staticmethod
     def query_and_cache(model, field, filter):
