+++ pipeline/tax_credit/logging_db_backend/base.py
@@ -8,9 +8,9 @@ logger = LoggerFactory.get(__name__)
 
 class DatabaseWrapper(PostgisDatabaseWrapper):
     def __init__(self, *args, **kwargs):
-        logger.info("Creating custom db wrapper")
+        logger.debug("Creating custom db wrapper")
         super().__init__(*args, **kwargs)
 
     def get_new_connection(self, conn_params):
-        logger.info("< ===== GETTING NEW CONNECTION ===== >")
+        logger.debug("< ===== GETTING NEW CONNECTION ===== >")
         return super().get_new_connection(conn_params)
