+++ pipeline/tax_credit/management/commands/load_assoc_table.py
@@ -1,254 +1,148 @@
-"""Loads tax credit programs for geographies into the database.
-"""
-import sys
-from itertools import islice
-
 from common.logger import LoggerFactory
 from django.conf import settings
 from django.core.management.base import BaseCommand, CommandParser
-from django.db import connections
-from django.db.models import Model
+from django.db import connection
+from django.db.models import F, Q, Subquery
 from django.db.models.functions import Substr
+from tax_credit.config_reader import get_load_config_reader
+from tax_credit.database_helper import DatabaseHelper
 from tax_credit.models import Geography, TargetBonusAssoc
 
 logger = LoggerFactory.get(__name__)
 
+config = get_load_config_reader(settings.CONFIG_FILE)
+
 
 class Command(BaseCommand):
-    """
-    Populates the `GeographyMetric` table.
+    def add_arguments(self, parser: CommandParser):
+        parser.add_argument(
+            "--target",
+            nargs="+",
+            default=[],
+            help="Restrict to the targets that will be used for associations",
+        )
+        parser.add_argument(
+            "--bonus",
+            nargs="+",
+            default=[],
+            help="Restrict to the bonuses that will be used for associations",
+        )
 
-    References:
-    - https://docs.djangoproject.com/en/4.1/howto/custom-management-commands/
-    - https://docs.djangoproject.com/en/4.1/topics/settings/
-    """
-
-    help = "Loads data to populate the geography metric table."
-
-    def add_arguments(self, parser: CommandParser) -> None:
-        """ """
-        parser.add_argument("--only-assoc", nargs="+", default=[])
-
-    def handle(self, *args, **options) -> None:
-        """
-        Executes the command. Accepts variable
-        numbers of keyword and non-keyword arguments.
-
-        Parameters:
-            None
-
-        Returns:
-            None
-        """
-
-        # Ugly code... essentially if the option --only assoc, it will load only the geography types passed. Otherwise it loads all geographies in succession
-        targets = ["state", "county", "municipal_util", "rural_coop"]
-        bonuses = ["distressed", "energy", "justice40", "low_income"]
-        logger.info("Building association table")
-        only_assoc = options["only_assoc"]
-        logger.info(f"Here is only_assoc : {only_assoc}")
-        if not only_assoc:
-            only_assoc.extend(targets)
-            only_assoc.extend(bonuses)
-
-        for target in [t for t in targets if t in only_assoc]:
-            for bonus in [b for b in bonuses if b in only_assoc]:
-                if target in ["state", "county"] and bonus in [
-                    "energy",
-                    "justice40",
-                    "low_income",
-                ]:
-                    if target == "state":
-                        self._find_and_load_matching_state_fips(target, bonus)
-                    elif target == "county":
-                        self._find_and_load_matching_county_fips(target, bonus)
-                    else:
-                        raise ValueError(
-                            "This shouldn't happen. Value was supposed to be in [ state, county ] ."
-                        )
-                else:
-                    logger.info("Need to use overlaps")
-                    self._find_and_load_overlaps(target, bonus)
-
-                # self.recycle_connection(Geography)
-                # self.recycle_connection(TargetBonusAssoc)
-
-        logger.info("Association table is finished")
-
-    def _find_and_load_overlaps(self, target_geom, bonus_geom):
-        # TODO dry this out and use this batching for all load patterns
-
-        logger.info(f"Finding overlaps between {target_geom} and {bonus_geom}")
-        for start in range(0, Geography.objects.count(), settings.SMALL_CHUNK_SIZE):
-            target_iter = Geography.objects.filter(geography_type__name=target_geom)[
-                start : start + settings.SMALL_CHUNK_SIZE
-            ]
-
-            self.recycle_connection(Geography)
-            for target in target_iter:
-                bonus_iter = (
-                    Geography.objects.filter(
-                        geography_type__name=bonus_geom,
-                        boundary__intersects=target.boundary,
-                    )
-                    .exclude(boundary__touches=target.boundary)
-                    .iterator()
+    def handle(self, *args, **options):
+        for job in config.assoc_jobs:
+            allowed_targets = options["target"]
+            if allowed_targets:
+                logger.info(f"Allowed targets {allowed_targets}")
+            if allowed_targets and job.target not in allowed_targets:
+                logger.info(
+                    f"SKIPPING association job [ {job.target} <-> {job.bonus} ], not in allowed targets"
                 )
+                continue
 
-                while True:
-                    try:
-                        batch = list(islice(bonus_iter, settings.SMALL_CHUNK_SIZE))
-                    except Exception as e:
-                        logger.error(f"Error with the batch : {e}")
-                        batch = []
-                    logger.info(f"Size of batch to load: {sys.getsizeof(batch)}")
-                    if not batch:
-                        logger.info(f"Loading finished for : {target_geom} {target}")
-                        break
-                    assocs = []
-                    for bonus in batch:
-                        assocs.append(
-                            TargetBonusAssoc(
-                                target_geography=target,
-                                target_geography_type=target.geography_type.name,
-                                bonus_geography=bonus,
-                                bonus_geography_type=bonus.geography_type.name,
-                            )
-                        )
-                    if assocs:
-                        logger.info(
-                            f"Loading batch of associations, {target_geom} to {bonus_geom} : {assocs}"
-                        )
-                        TargetBonusAssoc.objects.bulk_create(
-                            assocs,
-                            settings.SMALL_CHUNK_SIZE,
-                            update_conflicts=True,
-                            unique_fields=[
-                                "target_geography",
-                                "bonus_geography",
-                            ],
-                            update_fields=[
-                                "target_geography_type",
-                                "bonus_geography_type",
-                            ],
-                        )
-
-    def _find_and_load_matching_state_fips(self, target_geom, bonus_geom):
-        logger.info(f"Finding overlaps between {target_geom} and {bonus_geom}")
-        target_iter = Geography.objects.filter(
-            geography_type__name=target_geom
-        ).iterator(chunk_size=settings.SMALL_CHUNK_SIZE)
-
-        for target in target_iter:
-            assocs = []
-            bonus_iter = (
-                Geography.objects.annotate(state_fips=Substr("fips_info", 1, 2))
-                .filter(
-                    geography_type__name=bonus_geom,
-                    state_fips=target.fips_info,
+            allowed_bonuses = options["bonus"]
+            if allowed_bonuses:
+                logger.info(f"Allowed bonuses : {allowed_bonuses}")
+            if allowed_bonuses and job.bonus not in allowed_bonuses:
+                logger.info(
+                    f"SKIPPING association job [ {job.target} <-> {job.bonus} ], not in allowed bonuses"
                 )
-                .iterator()
+                continue
+
+            logger.info(f"Running job : {job.job_name}")
+
+            # cache target pk's
+            target_type_filter = Q(geography_type__name=job.target)
+            target_records: list = DatabaseHelper.query_and_cache(
+                Geography, "id", target_type_filter
             )
-            for bonus in bonus_iter:
-                assocs.append(
-                    TargetBonusAssoc(
-                        target_geography=target,
-                        target_geography_type=target.geography_type.name,
-                        bonus_geography=bonus,
-                        bonus_geography_type=bonus.geography_type.name,
-                    )
+
+            for target in target_records:
+                target_record = Geography.objects.filter(id=target.id).values(
+                    "id", "fips", "geography_type__name"
+                )[:1]
+
+                strategy = job.assoc_strategy.lower()
+                matches = self.match_fips(strategy, target_record, job.bonus)
+
+                matches = matches.annotate(
+                    target_id=target_record.values("id"),
+                    target_geography_type_name=target_record.values(
+                        "geography_type__name"
+                    ),
+                ).values(
+                    bonus_geography_type=F("geography_type__name"),
+                    bonus_geography_id=F("id"),
+                    target_geography_type=F("target_geography_type_name"),
+                    target_geography_id=F("target_id"),
                 )
-
-            while True:
-                batch = list(islice(bonus_iter, settings.SMALL_CHUNK_SIZE))
-                logger.info(f"Size of batch to load: {sys.getsizeof(batch)}")
-                if not batch:
-                    logger.info(f"Loading finished for : {target_geom} {target}")
-                    break
-                assocs = []
-                for bonus in batch:
-                    assocs.append(
-                        TargetBonusAssoc(
-                            target_geography=target,
-                            target_geography_type=target.geography_type.name,
-                            bonus_geography=bonus,
-                            bonus_geography_type=bonus.geography_type.name,
-                        )
-                    )
-                if assocs:
-                    logger.info(
-                        f"Loading batch of associations, {target_geom} to {bonus_geom} : {assocs}"
-                    )
-                    TargetBonusAssoc.objects.bulk_create(
-                        assocs,
-                        update_conflicts=True,
-                        unique_fields=["target_geography", "bonus_geography"],
-                        update_fields=[
-                            "target_geography_type",
-                            "bonus_geography_type",
-                        ],
-                    )
-
-    def _find_and_load_matching_county_fips(self, target_geom, bonus_geom):
-        logger.info(f"Finding overlaps between {target_geom} and {bonus_geom}")
-        target_iter = Geography.objects.filter(
-            geography_type__name=target_geom
-        ).iterator(chunk_size=settings.SMALL_CHUNK_SIZE)
-
-        for target in target_iter:
-            logger.info(f"\nTarget : {target.fips_info}")
-            assocs = []
-            bonus_iter = Geography.objects.filter(
-                geography_type__name=bonus_geom,
-                fips_info=target.fips_info,
-            ).iterator()
-            for bonus in bonus_iter:
-                assocs.append(
-                    TargetBonusAssoc(
-                        target_geography=target,
-                        target_geography_type=target.geography_type.name,
-                        bonus_geography=bonus,
-                        bonus_geography_type=bonus.geography_type.name,
-                    )
+                raw_sql_query, params = matches.query.sql_with_params()
+                insert_query = (
+                    f"INSERT INTO {TargetBonusAssoc._meta.db_table} "
+                    "("
+                    "target_geography_type, "
+                    "bonus_geography_type, "
+                    "bonus_geography_id, "
+                    "target_geography_id"
+                    ") "
+                    "SELECT "
+                    "target_geography_type, "
+                    "bonus_geography_type, "
+                    "bonus_geography_id, "
+                    "target_geography_id "
+                    f"FROM ({raw_sql_query}) AS subquery "
+                    "ON CONFLICT DO NOTHING"
                 )
-            # logger.info(f"Finished iterating bonuses, assocs : {assocs}")
-            while True:
-                batch = list(islice(bonus_iter, settings.SMALL_CHUNK_SIZE))
-                logger.info(f"Size of batch to load: {sys.getsizeof(batch)}")
-                if not batch:
-                    logger.info(f"Loading finished for : {target_geom} {target}")
-                    break
-                assocs = []
-                for bonus in batch:
-                    assocs.append(
-                        TargetBonusAssoc(
-                            target_geography=target,
-                            target_geography_type=target.geography_type.name,
-                            bonus_geography=bonus,
-                            bonus_geography_type=bonus.geography_type.name,
-                        )
+                with connection.cursor() as cursor:
+                    cursor.execute(
+                        insert_query,
+                        params,
                     )
-                if assocs:
-                    logger.info(
-                        f"Loading batch of associations, {target_geom} to {bonus_geom} : {assocs}"
-                    )
-                    TargetBonusAssoc.objects.bulk_create(
-                        assocs,
-                        update_conflicts=True,
-                        unique_fields=["target_geography", "bonus_geography"],
-                        update_fields=[
-                            "target_geography_type",
-                            "bonus_geography_type",
-                        ],
-                    )
+    
+    def match_fips(self, strategy: str, target_record, bonus: str):
+        """Matches geographies to the entities that contain them using one of several defined strategies
 
-    @staticmethod
-    def recycle_connection(model: Model):
-        """Recycles the connection from the model from the previous job.
-        Without this the connection will expire and cause SSL errors when working with Neon Postgres.
+        Args:
+            strategy (str): The name of the matching strategy to use
+            target_record: The target record to match on
+            bonus (str): The bonus type to match on
+
+        Raises:
+            ValueError: If the matching strategy is not defined, raises
         """
-        db_string = model.objects.db
-        logger.info(f"Recycling connection : {model}, {db_string}")
-        conn = connections[db_string]
-        if conn:
-            conn.close()
+        # TODO validate valid fips
+        if strategy == "fips_county":
+            matches = self.match_county_fips(
+                target_record.values("fips")[:5], bonus
+            )
+        elif strategy == "fips_state":
+            matches = self.match_state_fips(
+                target_record.values("fips")[:2], bonus
+            )
+        elif strategy == "overlap":
+            matches = self.match_overlap(target_record.values("id"), bonus)
+        else:
+            raise ValueError(
+                f"Invalid association strategy : {strategy}"
+            )
+        return matches
+
+    def match_county_fips(self, fips, bonus: str):
+        return Geography.objects.annotate(county_fips=Substr("fips", 1, 5)).filter(
+            geography_type__name=str(bonus), county_fips=fips
+        )
+
+    def match_state_fips(self, fips, bonus: str):
+        # It may seem a little funny matching a substring to a full fips string
+        # but remember that state fips info will only have 2 characters to begin
+        # with
+        return Geography.objects.annotate(state_fips=Substr("fips", 1, 2)).filter(
+            geography_type__name=bonus,
+            state_fips=fips,
+        )
+
+    def match_overlap(self, target_id: str, bonus_type: str):
+        target_boundary = Geography.objects.filter(id=target_id).values("geometry")
+        return Geography.objects.filter(
+            geography_type__name=bonus_type,
+            geometry__intersects=Subquery(target_boundary),
+        )
