+++ pipeline/tax_credit/management/commands/load_dependent_tables.py
@@ -30,6 +30,6 @@ class Command(BaseCommand):
                 )
             except RuntimeError as re:
                 logger.error(
-                    "ERROR!!!!! Could not validate load job. Records will not load."
+                    f"ERROR!!!!! Could not validate load job [ {job.job_name} ]. Records will not load."
                 )
                 logger.error(f"ERROR : {re}")
