+++ pipeline/tax_credit/models.py
@@ -20,15 +20,19 @@ class Geography(models.Model):
     # Autogenerate id field
     id = models.BigAutoField(primary_key=True)
     name = models.CharField(max_length=255)
+    fips = models.CharField(max_length=255, null=True)
     geography_type = models.ForeignKey(GeographyType, on_delete=models.CASCADE)
-    boundary = MultiPolygonField()
     as_of = models.DateField()
+    published_on = models.DateField(null=True)
     source = models.CharField(max_length=255)
-    fips_info = models.CharField(max_length=255, null=True)
+    geometry = MultiPolygonField()
 
     class Meta:
         db_table = "tax_credit_geography"
-        unique_together = [["name", "geography_type"]]
+        unique_together = [["name", "geography_type", "fips"]]
+    
+    def __str__(self):
+        return f"[ {self.id}, {self.name}, {self.fips}, {self.geography_type} ]"
 
 
 class Program(models.Model):
@@ -69,19 +73,11 @@ class TargetBonusAssoc(models.Model):
         return f"{self.target_geography.name} <-> {self.bonus_geography.name}"
 
 
-class CensusTract(models.Model):
-    id = models.CharField(max_length=255, primary_key=True)
-    centroid = PointField()
-    population = models.IntegerField()
-
-    class Meta:
-        db_table = "tax_credit_census_tract"
-
-
 class CensusBlockGroup(models.Model):
     id = models.CharField(max_length=255, primary_key=True)
     centroid = PointField()
     population = models.IntegerField()
+    year = models.IntegerField()
 
     class Meta:
         db_table = "tax_credit_census_block_group"
